backend_service/requirements.txt
mysql~=0.0.3
Django~=5.0.3
djangorestframework~=3.14.0